+++ NEWS.md
@@ -1,3 +1,7 @@
+# cffr (development version)
+
+- Add DOI <https://doi.org/10.5281/zenodo.5509766>
+
 # cffr 0.0.1
 
 * First stable release